requirements.txt
asgiref==3.5.0
Django==4.0.1
sqlparse==0.4.2
django-environ==0.8.1
djangorestframework==3.13.1
markdown==3.3.6
django-filter==21.1
djangorestframework-simplejwt==5.0.0
psycopg2==2.8.5
